refactor(react): move rendering into utility function
