+++ packages/react/lib/fragments.js
@@ -1,0 +1,13 @@
+const { renderToString } = require('react-dom/server');
+const {
+  Helmet: { renderStatic },
+} = require('react-helmet');
+
+module.exports = (element) => {
+  const reactMarkup = renderToString(element);
+  const fragments = Object.entries(renderStatic()).reduce(
+    (result, [key, value]) => ({ ...result, [key]: value.toString() }),
+    { reactMarkup, headPrefix: '', headSuffix: '' }
+  );
+  return fragments;
+};
